Renamed utils/utils.ts to just be utils/index.ts to imply a different meaning.
Added a toJson function.
