+++ src/permission_manager/permissionManager.service.ts
@@ -3,7 +3,7 @@ import { pino } from "pino";
 import { singleton } from "tsyringe";
 
 import { createLogger } from "../utils/logger/logger.decorator.js";
-import { deepMerge } from "../utils/utils.js";
+import { deepMerge } from "../utils/index.js";
 import { PermissionKeys } from "./addCommandKeys.decorator.js";
 import { Permission, PermissionKeysType, PermissionManagerConfig, PermissionMode } from "./models/index.js";
 import { PermissionManagerRepository } from "./permissionManager.repository.js";
