+++ src/role_manager/roleManager.service.ts
@@ -5,7 +5,7 @@ import { singleton } from "tsyringe";
 import { createLogger } from "../utils/logger/logger.decorator.js";
 import { Client } from "../utils/models/client.js";
 import { Task } from "../utils/models/task.js";
-import { fetchMessages } from "../utils/utils.js";
+import { fetchMessages } from "../utils/index.js";
 import { RoleManagerConfig } from "./roleManager.model.js";
 import { RoleManagerRepository } from "./roleManager.repository.js";
