+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 import { Client, Message, Snowflake, TextChannel } from "discord.js";
+import { ToJsonBase } from "./objects/toJsonBase.js";
 
 export async function fetchMessages(client: Client, channelId: Snowflake, messageIds: Snowflake[]): Promise<Message[]> {
     const result: Message[] = [];
@@ -15,6 +16,29 @@ export async function fetchMessages(client: Client, channelId: Snowflake, messag
     return result;
 }
 
+/**
+ * Wrapper function for JSON string conversation.
+ * Object returned will be the assigned object.
+ * If object extends type ToJsonBase then the fromJson function will be called.
+ * @param t Object to be assigned to.
+ * @param str Json string.
+ * @return Newly created object.
+ */
+export function toJson<T>(t: T, str: string): T {
+    if (t instanceof ToJsonBase) {
+        return t.fromJson(str) as unknown as T;
+    }
+
+    return Object.assign(t, JSON.parse(str));
+}
+
+/**
+ * Performs a deep merge on an object.
+ * If the object extends SanitizeObjectBase then the sanitize function will be called.
+ * @param target Target object.
+ * @param sources All sources to merge from.
+ * @return The newly created object.
+ */
 export function deepMerge<T, O>(target: T, ...sources: O[]): T {
     if (!sources.length) return target;
